BoxCollisionを修正中

+++ Hack_and_Slash/Actor.cpp
@@ -1,28 +1,28 @@
 #include "Actor.hpp"
 
 //コンストラクタ
-Actor::Actor(Entry* e,glm::ivec2 pos ,glm::ivec2 vec )
+Actor::Actor(Entry* e,glm::vec2 pos ,glm::vec2 vec )
 {
 	vector = vec;
 	position = pos;
-	size = glm::ivec2(0,0);
+	size = glm::vec2(0,0);
 	owner = e;
 }
 
 //方向を取得
-glm::ivec2 Actor::getVector()
+glm::vec2 Actor::getVector()
 {
 	return vector;
 }
 
 //座標を取得
-glm::ivec2 Actor::getPosition()
+glm::vec2 Actor::getPosition()
 {
 	return position;
 }
 
 //座標を取得
-void Actor::setPosition(glm::ivec2 pos)
+void Actor::setPosition(glm::vec2 pos)
 {
 	position = pos;
 }
