+++ Hack_and_Slash/Actor.hpp
@@ -16,24 +16,24 @@ class Entry;
 class Actor
 {
 public:
-	Actor(Entry* e,glm::ivec2 pos = glm::ivec2(0, 0), glm::ivec2 vec = glm::ivec2(0, 0));	//コンストラクタ
+	Actor(Entry* e,glm::vec2 pos = glm::vec2(0, 0), glm::vec2 vec = glm::vec2(0, 0));	//コンストラクタ
 	~Actor();	//デストラクタ
 
 	void virtual Update() = 0;	//計算
 	void virtual Draw() = 0;	//描画
 
 	//取得　関係
-	glm::ivec2 getVector();		//方向
-	glm::ivec2 getPosition();	//座標
+	glm::vec2 getVector();		//方向
+	glm::vec2 getPosition();	//座標
 
 	//設定　関係
-	void setPosition(glm::ivec2 pos);
+	void setPosition(glm::vec2 pos);
 
 protected:
 
-	glm::ivec2 position;	//座標
-	glm::ivec2 vector;		//方向
-	glm::ivec2 size;		//スプライトのサイズ
+	glm::vec2 position;	//座標
+	glm::vec2 vector;	//方向
+	glm::vec2 size;		//スプライトのサイズ
 
 	Entry* owner;			//Entry クラス
